Refactoring config util file
